requirements.txt
astroid==3.3.5
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dill==0.3.9
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
generateApiKey==0.4.7
greenlet==3.1.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
mysqlclient==2.2.5
platformdirs==4.3.6
pycparser==2.22
pylint==3.3.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
tomlkit==0.13.2
typing_extensions==4.12.2
validators==0.34.0
Werkzeug==3.0.4
